cdk_display_pipeline/lambda/format_image/requirements.txt
pillow==10.4.0
pillow-heif==0.18.0
